fix(core): move appendExtension to backfill script to not change any production code
Part of ENG-874
